W-16280695: Moved constant file to internal package and removed mule runtime properties from pom since it's already included as part of a parent module pom
