+++ src/main/java/org/mule/extension/ftp/internal/constant/FileDisplayConstants.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.ftp.api;
+package org.mule.extension.ftp.internal.constant;
 
 /**
  * Utility class to hold all the common constants between File related extensions for UI enhancement
